+++ PixLib OpenSource/NSDate+PixLib.h
@@ -35,7 +35,7 @@
 #define SECONDS_PER_YEAR 31536000
 
 typedef struct {
-    int year; int month; int day; int hour; int minute; int second;
+    NSInteger year; NSInteger month; NSInteger day; NSInteger hour; NSInteger minute; NSInteger second;
 } PxDateOptions;
 
 /**
